fix(mcp): rename clean-query-mcp in clean_query_mcp, entry point completo
La directory tools/clean-query-mcp/ (con trattino) non era importabile
come package Python. Rinomina e fix completi:

- git mv tools/clean-query-mcp/ -> tools/clean_query_mcp/
- server.py: from catalog -> from .catalog (import relativi)
- catalog.py: nessun cambio (usa solo lib standard + lab-connectors)
- Aggiunto __init__.py per rendere tools.clean_query_mcp un package
- pyproject.toml: packages include tools.*, [project.scripts] clean-query-mcp
- Rimosso requirements.txt (mcp ora in pyproject.toml)
- tests: import server -> from tools.clean_query_mcp import server
- tests: import catalog -> from tools.clean_query_mcp import catalog
- conftest.py: rimosso path hack per tools/ (non serve piu')
- CI workflows: aggiornati path e requirements
- README/CONTRIBUTING: aggiornati riferimenti
- opencode.json: command -> .venv/bin/clean-query-mcp
