--- tools/clean-query-mcp/requirements.txt
@@ -1,3 +1,0 @@
-mcp>=1.26.0
-duckdb==1.5.1
-git+https://github.com/dataciviclab/lab-connectors.git@v0.13.4
