+++ tools/clean_query_mcp/__init__.py
@@ -1,0 +1,1 @@
+"""Clean Query MCP: interrogazione dataset clean del DataCivicLab tramite DuckDB."""
