+++ tools/clean_query_mcp/catalog.py
