+++ tools/clean_query_mcp/scripts/smoke.py
